requirements.txt
opencv-python~=4.9.0.80
dlib~=19.24.6
torch~=2.3.1
numpy~=1.26.4
torchvision~=0.18.1
Pillow~=10.0.1
scikit-learn~=1.3.2
tqdm~=4.66.4
pytest~=8.3.3
pandas~=2.2.3
matplotlib~=3.9.2
mtcnn~=1.0.0
pathos~=0.3.2